rust: add a series of lints
Provide infrastructure for running lints across variable Rust versions.
The idea is that the lint groups (complexity, perf, style, suspicious)
are disabled, so that future changes to the compiler do not break.  These
are more frequent than changes to -Werror, and the Rust code will probably
evolve more quickly, so such breakage is expected to happen relatively
often.  An initial list of things that we _do_ want to keep an eye on,
because they are relatively common errors, is included in the workspace
Cargo.toml.

However, in order to keep up with the change, a CI job runs clippy with
Rust nightly, with failures marked as non-fatal.  Whenever such a failure
happens, the code should be fixed in QEMU and the lint added to the
already-long list in Cargo.toml (to ensure no regressions).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>